Make forms, functions and signals files splitable.
In order to make big files splitable for readability into a directory named after the
file type, the `import_signals_and_functions` function now check if the
module `{forms|functions|signals}` is a file or a directory.
If the module is a file, it is imported
If the module is a directory, all files in the directory are imported.

The function `import_mod` has been created to avoid `try ... except`
redundancy.